+++ Packer/Program.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021 Eberhard Beilharz
+// Copyright (c) 2017-2022 Eberhard Beilharz
 // This software is licensed under the GNU General Public License version 3
 // (https://opensource.org/licenses/GPL-3.0)
 
@@ -9,8 +9,8 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Security.Cryptography;
-using System.Text;
 using System.Text.RegularExpressions;
+using System.Threading.Tasks;
 using DigitaleBriefwahl;
 using DigitaleBriefwahl.Encryption;
 using DigitaleBriefwahl.ExceptionHandling;
@@ -24,12 +24,37 @@ namespace Packer
 	internal static class Program
 	{
 		private static Configuration Config { get; set; }
+		private static string Token { get; set; }
+		private static bool UseGist { get; set; } = true;
 		private static string ExecutableLocation => Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 
 		public static void Main(string[] args)
 		{
 			ExceptionLogging.Initialize("5012aef9a281f091c1fceea40c03003b", "Packer", args);
-			CreateBundle();
+
+			for (var i = 0; i < args.Length; i++)
+			{
+				var arg = args[i];
+				switch (arg)
+				{
+					case "--token":
+						if (i + 1 < args.Length)
+						{
+							Token = args[++i];
+						}
+						else
+							Help();
+						break;
+					case "--no-gist":
+						UseGist = false;
+						break;
+					default:
+						Help();
+						break;
+				}
+			}
+
+			Task.WaitAll(CreateBundle(UseGist, Token));
 
 			if (Debugger.IsAttached)
 				return;
@@ -38,12 +63,22 @@ namespace Packer
 			Console.ReadLine();
 		}
 
-		private static bool CreateBundle()
+		private static void Help()
+		{
+			Console.WriteLine("Usage:");
+			Console.WriteLine("packer [--token <GithubToken>] [--no-gist]");
+			Console.WriteLine();
+			Console.WriteLine("--token <GitHubToken>\tGitHub token used to authenticate");
+			Console.WriteLine("--no-gist\tManual upload to Google Drive or OneDrive instead of uploading to Gist");
+			Environment.Exit(1);
+		}
+
+		private static async Task<bool> CreateBundle(bool uploadToGist, string token)
 		{
 			if (!File.Exists(Configuration.ConfigName))
 			{
-				Console.Error.WriteLine();
-				Console.Error.WriteLine($"Config: The configuration file '{Configuration.ConfigName}' is missing. Exiting.");
+				await Console.Error.WriteLineAsync();
+				await Console.Error.WriteLineAsync($"Config: The configuration file '{Configuration.ConfigName}' is missing. Exiting.");
 				return false;
 			}
 
@@ -52,15 +87,15 @@ namespace Packer
 			var packCompiler = new PackCompiler(ExecutableLocation);
 			if (!File.Exists(packCompiler.ConfigFilename))
 			{
-				Console.Error.WriteLine();
-				Console.Error.WriteLine($"PackCompiler: The configuration file '{packCompiler.ConfigFilename}' is missing. Exiting.");
+				await Console.Error.WriteLineAsync();
+				await Console.Error.WriteLineAsync($"PackCompiler: The configuration file '{packCompiler.ConfigFilename}' is missing. Exiting.");
 				return false;
 			}
 
 			if (!File.Exists(packCompiler.Config.PublicKey))
 			{
-				Console.Error.WriteLine();
-				Console.Error.WriteLine(
+				await Console.Error.WriteLineAsync();
+				await Console.Error.WriteLineAsync(
 					$"PackCompiler: The public key file '{packCompiler.Config.PublicKey}' specified in {packCompiler.ConfigFilename} is missing. Exiting.");
 				return false;
 			}
@@ -71,30 +106,55 @@ namespace Packer
 			var zipFileHash = CalculateHash(zipFile);
 			Console.WriteLine();
 			Console.WriteLine("Files packed successfully.");
-			Console.WriteLine($"Now upload the file '{Path.GetFileName(zipFile)}' in directory");
-			Console.WriteLine($"'{Path.GetDirectoryName(zipFile)}'");
-			Console.WriteLine("to a suitable website.");
-			Console.WriteLine("Then enter the full Download URL for that file:");
-			Uri url = null;
-			for (var urlOk = false; !urlOk;)
+			string urlString = null;
+			if (uploadToGist)
 			{
-				Console.Write("> ");
-				var urlString = Console.ReadLine();
-				try
+				if (string.IsNullOrEmpty(token))
 				{
-					urlString = AdjustUrlIfGoogleDrive(urlString);
-					urlString = AdjustUrlIfMicrosoftOneDrive(urlString);
-					url = new Uri(urlString);
-					urlOk = TryDownload(url, zipFileHash);
+					Console.Write("Enter the GitHub personal access token for authentication: ");
+					token = Console.ReadLine();
 				}
-				catch (UriFormatException)
+
+				Console.WriteLine("Uploading to GitHub Gist...");
+				var gist = new GitHubGist();
+				Task.WaitAll(gist.Initialize(token));
+				urlString = await gist.Upload(packCompiler.Config.Title, zipFile);
+				if (!string.IsNullOrEmpty(urlString))
+					Console.WriteLine("Upload succeeded.");
+			}
+			else
+			{
+				Console.WriteLine(
+					$"Now upload the file '{Path.GetFileName(zipFile)}' in directory");
+				Console.WriteLine($"'{Path.GetDirectoryName(zipFile)}'");
+				Console.WriteLine("to a suitable website.");
+				Console.WriteLine("Then enter the full Download URL for that file:");
+				for (var urlOk = false; !urlOk;)
 				{
-					Console.Error.WriteLine("Invalid URL. Please enter a valid download URL:");
+					Console.Write("> ");
+					urlString = Console.ReadLine();
+					try
+					{
+						urlString = AdjustUrlIfGoogleDrive(urlString);
+						urlString = AdjustUrlIfMicrosoftOneDrive(urlString);
+						var url = new Uri(urlString);
+						urlOk = TryDownload(url, zipFileHash);
+						if (urlOk)
+							urlString = url.ToString();
+					}
+					catch (UriFormatException)
+					{
+						await Console.Error.WriteLineAsync(
+							"Invalid URL. Please enter a valid download URL:");
+					}
 				}
 			}
 
+			if (string.IsNullOrEmpty(urlString))
+				return false;
+
 			var urlFile = Path.ChangeExtension(zipFile, "wahlurl");
-			File.WriteAllText(urlFile, url.ToString());
+			File.WriteAllText(urlFile, urlString);
 
 			Console.WriteLine();
 			Console.WriteLine($"The following files were created in {Path.GetDirectoryName(zipFile)}:");